organization of the proj
