+++ client/package.json
@@ -23,6 +23,7 @@
     "@angular/platform-browser-dynamic": "^17.0.0",
     "@angular/router": "^17.0.0",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@ngneat/until-destroy": "^10.0.0",
     "bootstrap": "5.3",
     "joi": "^17.11.0",
     "lodash": "^4.17.21",
