+++ client/src/app/app.component.html
@@ -2,11 +2,11 @@
   <mat-toolbar-row>
     <div class="nav">
       <div class="left-side-bar">
-        <a href="/" mat-button> home </a>
-        <a href="/about" mat-button> about </a>
+        <a routerLink="/" mat-button> home </a>
+        <a routerLink="/about" mat-button> about </a>
       </div>
       <div class="right-side-bar">
-        <a href="/admin" mat-button> admin </a>
+        <a routerLink="/admin" mat-button> admin </a>
       </div>
     </div>
   </mat-toolbar-row>
